problem analysis
