+++ package.json
@@ -7,6 +7,8 @@
     "test-data": "yarn ts-node index.ts"
   },
   "dependencies": {
+    "@types/node": "^20.4.2",
+    "fs": "^0.0.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.1.3"
   }
